Consolidate lib/route-tree.ts into routes.ts
