--- src/lib/router.ts
@@ -1,6 +1,0 @@
-import { createHashRouter } from 'react-router';
-import routeTree from './route-tree';
-
-const router = createHashRouter(routeTree);
-
-export default router;
