+++ src/main.tsx
@@ -13,7 +13,7 @@ import * as ReactDOM from 'react-dom/client';
 import { RouterProvider } from 'react-router';
 
 import queryClient from './lib/query-client';
-import router from './lib/router';
+import router from './routes';
 
 /*
 |--------------------------------------------------------------------------
